pipeline: move Azure I/O into its own file
