+++ docs/api/toasty.pipeline.azure_io.assert_enabled.rst
@@ -1,0 +1,6 @@
+assert_enabled
+==============
+
+.. currentmodule:: toasty.pipeline.azure_io
+
+.. autofunction:: assert_enabled
